requirements/sample-requirements.txt
pytorch-lightning
torchmetrics
torchvision
gradio
hydra-core
omegaconf
clearml
dvc[ssh]
